# Faster Q5_K and Q6_K on Metal