Along the same lines as #2290. Here the speedup is not quite as large as for `Q4_K`, but still significant:

 | Model | Master | This PR | Speedup |
|--:|--:|--:|--:|
| Q5_K_S 7B | 26.2 | 22.8 | 14.9% |
|Q5_K_S 13B | 46.1 | 39.3 | 17.4% |
|Q5_K_S 33B | 115.5 | 97.0 | 19.1% |
|Q5_K_S 65B | 214.6 | 181.1 | 18.5% | 
| Q6_K 7B | 25.6 | 24.6 | 4.1% |
|Q6_K 13B | 46.1 | 44.3 | 4.1% |
|Q6_K 33B | 116.6 | 111.0 | 5.0% |

Table shows token generation time in ms/t on M2 Max with 30-core GPU. The system has 64 GB RAM and the 65B `Q6_K` model does not run successfully.